TOFTOF: Fix resolution panel
The interface of the QSpinBox has change, so setting of the limits
requires integer values, but the chopper speed is given as floating
point values. Therefor the type of the spin box is changed to
QDoubleSpinBox.

Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35670
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>

+++ nicos_mlz/toftof/gui/resolutionpanel.py
@@ -151,7 +151,7 @@ class ResolutionPanel(NicosWidget, Panel):
         else:
             self.buttonBox.rejected.connect(self.close)
 
-        self.speed.valueChanged['int'].connect(self.recalculate)
+        self.speed.valueChanged['double'].connect(self.recalculate)
         self.waveLength.valueChanged['double'].connect(self.recalculate)
         self.ratio.valueChanged['int'].connect(self.recalculate)
         self.slits.currentIndexChanged.connect(self.recalculate)
